- register `/auth-hub/send-message` only in AuthHub - add `x-appmixer-version-slack` when calling authhub - tests
